+++ zerver/lib/user_message.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import Iterable, List
 
 from django.db import connection
 from psycopg2.extras import execute_values
@@ -21,6 +21,25 @@ class UserMessageLite:
 
     def flags_list(self) -> List[str]:
         return UserMessage.flags_list_for_flags(self.flags)
+
+
+DEFAULT_HISTORICAL_FLAGS = UserMessage.flags.historical | UserMessage.flags.read
+
+
+def create_historical_user_messages(
+    *, user_id: int, message_ids: Iterable[int], flags: int = DEFAULT_HISTORICAL_FLAGS
+) -> None:
+    # Users can see and interact with messages sent to streams with
+    # public history for which they do not have a UserMessage because
+    # they were not a subscriber at the time the message was sent.
+    # In order to add emoji reactions or mutate message flags for
+    # those messages, we create UserMessage objects for those messages;
+    # these have the special historical flag which keeps track of the
+    # fact that the user did not receive the message at the time it was sent.
+    UserMessage.objects.bulk_create(
+        UserMessage(user_profile_id=user_id, message_id=message_id, flags=flags)
+        for message_id in message_ids
+    )
 
 
 def bulk_insert_ums(ums: List[UserMessageLite]) -> None:
